+++ src/main/java/lotto/domain/game/LottoSetting.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.game;
 
 public class LottoSetting {
