+++ src/main/java/lotto/domain/number/LottoNumberSet.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.number;
 
 import java.util.HashSet;
 import java.util.List;
