+++ src/main/java/lotto/domain/result/LottoIncomeRate.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.result;
 
 public class LottoIncomeRate {
