+++ src/main/java/lotto/domain/result/LottoResult.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.result;
 
 import java.util.Arrays;
