+++ src/test/java/lotto/domain/LottoNumberTest.java
@@ -1,6 +1,6 @@
-package lotto;
+package lotto.domain;
 
-import lotto.domain.LottoNumber;
+import lotto.domain.number.LottoNumber;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
