+++ src/test/java/lotto/domain/LottoTicketListTest.java
@@ -1,10 +1,10 @@
-package lotto;
+package lotto.domain;
 
 import java.util.Arrays;
-import lotto.domain.LottoResult;
-import lotto.domain.LottoTicket;
-import lotto.domain.LottoTicketList;
-import lotto.domain.LottoWinningNumberList;
+import lotto.domain.result.LottoResult;
+import lotto.domain.ticket.LottoTicket;
+import lotto.domain.ticket.LottoTicketList;
+import lotto.domain.ticket.LottoWinningNumberList;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
