+++ src/test/java/lotto/domain/LottoWinningNumbersTest.java
@@ -1,9 +1,9 @@
-package lotto;
+package lotto.domain;
 
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Stream;
-import lotto.domain.LottoWinningNumberList;
+import lotto.domain.ticket.LottoWinningNumberList;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
